# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Authentication API",
        "description": "API endpoints for user authentication",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/game-stats": {
            "get": {
                "tags": [
                    "Game Stats"
                ],
                "summary": "Get all game stats",
                "description": "List all game stats",
                "operationId": "f7e2f19475da271af39f5fccfae95e27",
                "responses": {
                    "200": {
                        "description": "List of game stats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "team_id": {
                                                "type": "integer"
                                            },
                                            "game_id": {
                                                "type": "integer"
                                            },
                                            "points": {
                                                "type": "integer"
                                            },
                                            "field_goals_attempted": {
                                                "type": "integer"
                                            },
                                            "field_goals_made": {
                                                "type": "integer"
                                            },
                                            "field_goal_percentage": {
                                                "type": "number"
                                            },
                                            "team": {
                                                "type": "object"
                                            },
                                            "game": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Game Stats"
                ],
                "summary": "Create a new game stat",
                "description": "Create a new game stat",
                "operationId": "179cd8d238fee0f769b589d37b0c3cdf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "team_id",
                                    "game_id"
                                ],
                                "properties": {
                                    "team_id": {
                                        "type": "integer"
                                    },
                                    "game_id": {
                                        "type": "integer"
                                    },
                                    "points": {
                                        "type": "integer"
                                    },
                                    "field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "field_goals_made": {
                                        "type": "integer"
                                    },
                                    "three_point_field_goals_made": {
                                        "type": "integer"
                                    },
                                    "three_point_field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "two_point_field_goals_made": {
                                        "type": "integer"
                                    },
                                    "two_point_field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "free_throws_made": {
                                        "type": "integer"
                                    },
                                    "free_throws_attempted": {
                                        "type": "integer"
                                    },
                                    "offensive_rebounds": {
                                        "type": "integer"
                                    },
                                    "defensive_rebounds": {
                                        "type": "integer"
                                    },
                                    "total_rebounds": {
                                        "type": "integer"
                                    },
                                    "assists": {
                                        "type": "integer"
                                    },
                                    "turnovers": {
                                        "type": "integer"
                                    },
                                    "steals": {
                                        "type": "integer"
                                    },
                                    "blocks": {
                                        "type": "integer"
                                    },
                                    "personal_fouls": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Game stat created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "game_stat": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/game-stats/{id}": {
            "get": {
                "tags": [
                    "Game Stats"
                ],
                "summary": "Get game stat details",
                "description": "Get a specific game stat",
                "operationId": "ce22be9ac20bd7f21f88c62dff04f16f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game Stat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game stat details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Game Stats"
                ],
                "summary": "Update game stat details",
                "description": "Update a game stat",
                "operationId": "1200d6d17b8e4fc3a26d45e845658523",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game Stat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "team_id",
                                    "game_id"
                                ],
                                "properties": {
                                    "team_id": {
                                        "type": "integer"
                                    },
                                    "game_id": {
                                        "type": "integer"
                                    },
                                    "points": {
                                        "type": "integer"
                                    },
                                    "field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "field_goals_made": {
                                        "type": "integer"
                                    },
                                    "three_point_field_goals_made": {
                                        "type": "integer"
                                    },
                                    "three_point_field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "two_point_field_goals_made": {
                                        "type": "integer"
                                    },
                                    "two_point_field_goals_attempted": {
                                        "type": "integer"
                                    },
                                    "free_throws_made": {
                                        "type": "integer"
                                    },
                                    "free_throws_attempted": {
                                        "type": "integer"
                                    },
                                    "offensive_rebounds": {
                                        "type": "integer"
                                    },
                                    "defensive_rebounds": {
                                        "type": "integer"
                                    },
                                    "total_rebounds": {
                                        "type": "integer"
                                    },
                                    "assists": {
                                        "type": "integer"
                                    },
                                    "turnovers": {
                                        "type": "integer"
                                    },
                                    "steals": {
                                        "type": "integer"
                                    },
                                    "blocks": {
                                        "type": "integer"
                                    },
                                    "personal_fouls": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Game stat updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Game Stats"
                ],
                "summary": "Delete a game stat",
                "description": "Delete a game stat",
                "operationId": "37bc3c039da3545e8e48145f1170d5b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Game Stat ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game stat deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}